Fix [filter cgi] and [filter value], as per bug #380.
I don't see any way the old code could have *not* resulted in a Perl error,
but I may be totally misunderstanding the intention. But according to my
reading of the docs ("good way to get a starting value for a filter set")
and the code, this is how this should work:

when CGI foo = bar:

[filter cgi]foo[/filter] => bar

If this is correct, I'll commit the change to 4.8 as well.

Thanks to Jonathan Clark <jonc@webmaint.net> for reporting.