Stop dying if there's a problem with FedExexpress lookups, which kept even ground rate lookups from being done during the first request when the Business::Fedex object is first created.
Instead, just log the error but continue on with possible ground lookup.

(This situation resulted in the oddity of the shipping rate displaying
as $0.00 on the checkout page, but still getting added into the total cost
and displaying in the shipping rate pulldown select box just fine.)

Also removed a duplicated debug statement and one left over from the
ups_query tag, and prettified some default values initialization.