Add entirely new way to run Interchange: inside Apache/mod_perl, with no separate Interchange daemon and no link program. Still experimental and under development, but does work with a few exceptions.
POD documentation in scripts/ic_mod_perl.PL should be enough to get you
started if you want to try it, and details some of my rationale and the
pros and cons of doing this.

A key feature is that the exact same Interchange installation can be set
up to work either with the standalone IC daemon, or inside mod_perl, and
you can switch back and forth between the two with no code or config
changes.