Fix bug #387, problem with [bounce] target coming from [history-scan] when using browser's Back button:
(begin)
then click hammers.... then you'll see a hammer with no for stock... click
no... then once the submit form for letting a person know when it is in
stock comes up... click your browser's back button without filling out
any of the form... there is not even a button that I saw to click back...

then add some quantity in any of the items that are in stock... click
the buy list button and nothing will happen. nothing get put in your
cart... unless you do a refresh... so, the buy list button has a problem.
(end)