Fix mv_ip setting when using [item-list reverse=1].
I'd like to generalize this to allow iterate_*_list to accept a reverse=1
option which would have them walk the array in reverse order (rather than
first reversing the array, then handing it off, with all the speed loss),
and I think it would be a useful general array-walking feature.

However, iterate_hash_list alone has 3 different for loops which would
each have to be changed, and I can't think of a clean way to do either/or
for loops at the moment. Comments welcome.